backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic[email]
motor
pymongo
passlib[bcrypt]
python-jose[cryptography]
